requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
pillow
pytesseract
pydantic
requests
